Refactor program.cs and GeneralStartupExtensions a bit
