+++ DotNet/DataAcquisition/Program.cs
@@ -1,84 +1,34 @@
 using DataAcquisition.Domain.Application.Serializers;
-using HealthChecks.UI.Client;
 using Hl7.Fhir.Model;
 using Hl7.Fhir.Serialization;
 using LantanaGroup.Link.DataAcquisition.Domain.Application.Interfaces;
 using LantanaGroup.Link.DataAcquisition.Domain.Application.Serializers;
 using LantanaGroup.Link.DataAcquisition.Domain.Application.Services;
 using LantanaGroup.Link.DataAcquisition.Domain.Extensions;
-using LantanaGroup.Link.DataAcquisition.Domain.Infrastructure.Context;
+
 using LantanaGroup.Link.DataAcquisition.Domain.Settings;
 using LantanaGroup.Link.DataAcquisition.Jobs;
 using LantanaGroup.Link.DataAcquisition.Listeners;
-using LantanaGroup.Link.Shared.Application.Extensions;
 using LantanaGroup.Link.Shared.Application.Extensions.Security;
 using LantanaGroup.Link.Shared.Application.Factories;
-using LantanaGroup.Link.Shared.Application.Health;
 using LantanaGroup.Link.Shared.Application.Interfaces;
 using LantanaGroup.Link.Shared.Application.Middleware;
-using LantanaGroup.Link.Shared.Application.Models;
-using LantanaGroup.Link.Shared.Application.Models.Configs;
-using LantanaGroup.Link.Shared.Domain.Repositories.Interceptors;
 using LantanaGroup.Link.Shared.Settings;
-using Microsoft.AspNetCore.Authentication.JwtBearer;
-using Microsoft.AspNetCore.Diagnostics.HealthChecks;
 using Microsoft.EntityFrameworkCore;
-using Microsoft.OpenApi.Models;
-using System.Reflection;
 using System.Text.Json.Serialization;
 
 var builder = WebApplication.CreateBuilder(args);
 
-RegisterServices(builder);
-var app = builder.Build();
-SetupMiddleware(app);
+builder.RegisterAll(DataAcquisitionConstants.ServiceName, true);
 
-app.Run();
+builder.RegisterQuartzAcquisitionJob(builder.Configuration.GetConnectionString(ConfigurationConstants.DatabaseConnections.DatabaseConnection));
 
-#region Register Services
-
-static void RegisterServices(WebApplicationBuilder builder)
-{
-    var consumerSettings = builder.Configuration.GetRequiredSection(nameof(ConsumerSettings)).Get<ConsumerSettings>();
-
-    builder.Services.AddDbContext<DataAcquisitionDbContext>((sp, options) => {
-
-        var updateBaseEntityInterceptor = sp.GetRequiredService<UpdateBaseEntityInterceptor>();
-        var dbProvider =
-            builder.Configuration.GetValue<string>(DataAcquisitionConstants.AppSettingsSectionNames.DatabaseProvider);
-
-        switch (dbProvider)
-        {
-            case ConfigurationConstants.AppSettings.SqlServerDatabaseProvider:
-                string? connectionString = builder.Configuration.GetConnectionString(ConfigurationConstants.DatabaseConnections.DatabaseConnection);
-
-                if (string.IsNullOrEmpty(connectionString))
-                    throw new InvalidOperationException("Database connection string is null or empty.");
-
-                options
-                    .UseSqlServer(connectionString)
-                    .AddInterceptors(updateBaseEntityInterceptor);
-
-                break;
-            default:
-                throw new InvalidOperationException("Database provider not supported.");
-        }
-    });
-
-    builder.Services.AddScoped<DbContext, DataAcquisitionDbContext>();
-
-    builder.RegisterAll(DataAcquisitionConstants.ServiceName, true);
-
-    builder.Services.AddTransient<IRetryEntityFactory, RetryEntityFactory>();
-
-    builder.RegisterQuartzAcquisitionJob(builder.Configuration.GetConnectionString(ConfigurationConstants.DatabaseConnections.DatabaseConnection));
-
-    // Add services to the container.
-    // Additional configuration is required to successfully run gRPC on macOS.
-    // For instructions on how to configure Kestrel and gRPC clients on macOS, visit https://go.microsoft.com/fwlink/?linkid=2099682
-    builder.Services.AddControllers(
-        options => options.SuppressImplicitRequiredAttributeForNonNullableReferenceTypes = true
-        ).AddJsonOptions(options =>
+// Add services to the container.
+// Additional configuration is required to successfully run gRPC on macOS.
+// For instructions on how to configure Kestrel and gRPC clients on macOS, visit https://go.microsoft.com/fwlink/?linkid=2099682
+builder.Services.AddControllers(
+    options => options.SuppressImplicitRequiredAttributeForNonNullableReferenceTypes = true
+    ).AddJsonOptions(options =>
     {
         options.JsonSerializerOptions.DefaultIgnoreCondition = JsonIgnoreCondition.WhenWritingNull;
         options.JsonSerializerOptions.PropertyNameCaseInsensitive = true;
@@ -88,130 +38,55 @@ static void RegisterServices(WebApplicationBuilder builder)
         options.JsonSerializerOptions.Converters.Add(new QueryPlanPutModelConverter());
         options.JsonSerializerOptions.Converters.Add(new TimeSpanConverter());
         options.JsonSerializerOptions.ForFhir(ModelInfo.ModelInspector);
-    }); 
+    });
 
-    //Add Hosted Services
-    if (!consumerSettings?.DisableConsumer ?? true)
+builder.RegisterCors();
+builder.RegisterHealthChecks();
+builder.RegisterSwagger(includeAuth: true);
+// Add Link Security
+bool allowAnonymousAccess = builder.Configuration.GetValue<bool>("Authentication:EnableAnonymousAccess");
+builder.Services.AddLinkBearerServiceAuthentication(options =>
+{
+    options.Environment = builder.Environment;
+    options.AllowAnonymous = allowAnonymousAccess;
+    options.Authority = builder.Configuration.GetValue<string>("Authentication:Schemas:LinkBearer:Authority");
+    options.ValidateToken = builder.Configuration.GetValue<bool>("Authentication:Schemas:LinkBearer:ValidateToken");
+    options.ProtectKey = builder.Configuration.GetValue<bool>("DataProtection:Enabled");
+    options.SigningKey = builder.Configuration.GetValue<string>("LinkTokenService:SigningKey");
+});
+
+builder.Services.AddTransient<IRetryEntityFactory, RetryEntityFactory>();
+
+builder.RegisterHostedServices((services, settings) =>
+{
+    if (!settings?.DisableConsumer ?? true)
     {
-        builder.Services.AddHostedService<DataAcquisitionRequestedListener>();
-        builder.Services.AddHostedService<PatientCensusScheduledListener>();
+        services.AddHostedService<DataAcquisitionRequestedListener>();
+        services.AddHostedService<PatientCensusScheduledListener>();
     }
 
-    if (!consumerSettings?.DisableRetryConsumer ?? true)
+    if (!settings?.DisableRetryConsumer ?? true)
     {
         // TODO: Retry consumer services temporarily disabled for LNK-4038
-        //builder.Services.AddSingleton(new RetryListenerSettings(DataAcquisitionConstants.ServiceName, [KafkaTopic.DataAcquisitionRequestedRetry.GetStringValue(), KafkaTopic.PatientCensusScheduledRetry.GetStringValue()]));
-        //builder.Services.AddHostedService<RetryListener>();
-        //builder.Services.AddHostedService<RetryScheduleService>();
+        //services.AddSingleton(new RetryListenerSettings(DataAcquisitionConstants.ServiceName, [KafkaTopic.DataAcquisitionRequestedRetry.GetStringValue(), KafkaTopic.PatientCensusScheduledRetry.GetStringValue()]));
+        //services.AddHostedService<RetryListener>();
+        //services.AddHostedService<RetryScheduleService>();
     }
+});
 
-    // Add Link Security
-    bool allowAnonymousAccess = builder.Configuration.GetValue<bool>("Authentication:EnableAnonymousAccess");
-    builder.Services.AddLinkBearerServiceAuthentication(options =>
-    {
-        options.Environment = builder.Environment;
-        options.AllowAnonymous = allowAnonymousAccess;
-        options.Authority = builder.Configuration.GetValue<string>("Authentication:Schemas:LinkBearer:Authority");
-        options.ValidateToken = builder.Configuration.GetValue<bool>("Authentication:Schemas:LinkBearer:ValidateToken");
-        options.ProtectKey = builder.Configuration.GetValue<bool>("DataProtection:Enabled");
-        options.SigningKey = builder.Configuration.GetValue<string>("LinkTokenService:SigningKey");
-    });
+builder.Services.AddSingleton(TimeProvider.System);
+builder.Services.AddSingleton<IDataAcquisitionServiceMetrics, DataAcquisitionServiceMetrics>();
 
-    builder.Services.AddEndpointsApiExplorer();
-    builder.Services.AddSwaggerGen(c =>
-    {
-        if (!allowAnonymousAccess)
-        {
-            #region Authentication Schemas
-            c.AddSecurityDefinition("Bearer", new OpenApiSecurityScheme
-            {
-                Description = $"Authorization using JWT",
-                Name = "Authorization",
-                Type = SecuritySchemeType.Http,
-                BearerFormat = "JWT",
-                In = ParameterLocation.Header,
-                Scheme = JwtBearerDefaults.AuthenticationScheme
-            });
+var app = builder.Build();
 
-            c.AddSecurityRequirement(new OpenApiSecurityRequirement
-            {
-                {
-                    new OpenApiSecurityScheme
-                    {
-                        Reference = new OpenApiReference
-                        {
-                            Id = "Bearer",
-                            Type = ReferenceType.SecurityScheme
-                        }
-                    },
-                    new List<string>()
-                }
-            });
-            #endregion
-        }
+app.ConfigureCommonMiddleware("data");
 
-        var xmlFile = $"{Assembly.GetExecutingAssembly().GetName().Name}.xml";
-        var xmlPath = Path.Combine(AppContext.BaseDirectory, xmlFile);
-        c.IncludeXmlComments(xmlPath);
-        c.DocumentFilter<HealthChecksFilter>();
-    });    
-
-    //Add CORS
-    builder.Services.AddLinkCorsService(options => {
-        options.Environment = builder.Environment;
-    });
-
-    //Add Health Check
-    var kafkaConnection = builder.Configuration.GetRequiredSection(KafkaConstants.SectionName).Get<KafkaConnection>();
-    var kafkaHealthOptions = new KafkaHealthCheckConfiguration(kafkaConnection, DataAcquisitionConstants.ServiceName).GetHealthCheckOptions();
-
-    builder.Services.AddHealthChecks()
-        .AddDbContextCheck<DataAcquisitionDbContext>(HealthCheckType.Database.ToString())
-        .AddKafka(kafkaHealthOptions, HealthCheckType.Kafka.ToString());
-
-    builder.Services.AddSingleton(TimeProvider.System);
-    builder.Services.AddSingleton<IDataAcquisitionServiceMetrics, DataAcquisitionServiceMetrics>();
-}
-
-#endregion
-
-#region Set up middleware
-
-static void SetupMiddleware(WebApplication app)
+//check for anonymous access
+if (!allowAnonymousAccess)
 {
-    app.ConfigureSwagger();
-
-    app.AutoMigrateEF<DataAcquisitionDbContext>();
-
-    if (app.Environment.IsDevelopment())
-    {
-        app.UseDeveloperExceptionPage();
-    }
-    else
-    {
-        app.UseExceptionHandler();
-    }
-
-    app.UseRouting();
-    app.UseCors(CorsSettings.DefaultCorsPolicyName);
-
-    //check for anonymous access
-    var allowAnonymousAccess = app.Configuration.GetValue<bool>("Authentication:EnableAnonymousAccess");
-    if (!allowAnonymousAccess)
-    {
-        app.UseAuthentication();
-        app.UseMiddleware<UserScopeMiddleware>();
-    }
-    app.UseAuthorization();
-
-    app.MapControllers();
-
-    //map health check middleware and info endpoint
-    app.MapHealthChecks("/health", new HealthCheckOptions
-    {
-        ResponseWriter = UIResponseWriter.WriteHealthCheckUIResponse,
-    });
-    app.MapInfo(Assembly.GetExecutingAssembly(), app.Configuration, "data");
+    app.UseAuthentication();
+    app.UseMiddleware<UserScopeMiddleware>();
 }
+app.UseAuthorization();
 
-#endregion
+app.Run();
